requirements.txt
airtest
pocoui
airtest-selenium